Change the messaging for invalid Python versions
Sublime build 4201 and upwards redact away the Python 3.8 plugin host
and replace it with a Python 3.13 plugin host (possibly Python 3.14 may
be used in its place before the private beta ends).

In prior builds, a .python-version file that specifies an invalid
version of Python (i.e. not one of the versions that there are plugin
hosts for) would cause Sublime to reject the plugins in that package
and not load them, and the previous release included a change to
reflect this.

In the newer builds (still in private beta), this is not the case and
such plugins seem to load in the new 3.13 plugin host. It is unclear at
the moment if this is intended (most likely it is) since there are now
many packages that reference Python 3.8 that will break.

For the time being the verbiage used in the hover popup is changed to
indicate that plugins MAY be ignored, so that people running the non
beta builds and beta build users alike get the same messaging.